+++ src/main/kotlin/org/vaccineimpact/api/models/expectations/ExpectationMapping.kt
@@ -1,4 +1,4 @@
-package org.vaccineimpact.api.models
+package org.vaccineimpact.api.models.expectations
 
 data class ExpectationMapping(
         override val expectation: CountryOutcomeExpectations,
