+++ src/main/kotlin/org/vaccineimpact/api/models/expectations/OutcomeExpectations.kt
@@ -1,4 +1,6 @@
-package org.vaccineimpact.api.models
+package org.vaccineimpact.api.models.expectations
+
+import org.vaccineimpact.api.models.Outcome
 
 data class OutcomeExpectations(override val id: Int,
                                override val description: String,
