+++ src/main/kotlin/org/vaccineimpact/api/models/expectations/TouchstoneModelExpectations.kt
@@ -1,4 +1,4 @@
-package org.vaccineimpact.api.models
+package org.vaccineimpact.api.models.expectations
 
 import java.beans.ConstructorProperties
