+++ src/main/kotlin/org/vaccineimpact/api/models/responsibilities/ResponsibilityDetails.kt
@@ -1,6 +1,6 @@
 package org.vaccineimpact.api.models.responsibilities
 
-import org.vaccineimpact.api.models.CountryOutcomeExpectations
+import org.vaccineimpact.api.models.expectations.CountryOutcomeExpectations
 import org.vaccineimpact.api.models.TouchstoneVersion
 
 data class ResponsibilityDetails(
